project/requirements/remote.txt
-r common.txt

# server
uWSGI==2.0.26

# Prometheus Monitoring
prometheus-client==0.20.0
django-prometheus==2.3.1

# Sentry Monitoring
sentry-sdk==2.14.0

# Unittest Reporting
unittest-xml-reporting==3.2.0
